# schema/openapi.yaml
openapi: 3.1.0
info:
  title: AIRAS API
  version: 0.0.1
paths:
...
  /airas/v1/experiments/run-ids:
    post:
      tags:
      - experiments
      summary: Fetch Run Ids
      operationId: fetch_run_ids_airas_v1_experiments_run_ids_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FetchRunIdsRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchRunIdsResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/experiments/results:
    post:
      tags:
      - experiments
      summary: Fetch Experimental Results
      operationId: fetch_experimental_results_airas_v1_experiments_results_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FetchExperimentalResultsRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchExperimentalResultsResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/experiments/sanity-checks/dispatch:
    post:
      tags:
      - experiments
      summary: Dispatch Sanity Check
      operationId: dispatch_sanity_check_airas_v1_experiments_sanity_checks_dispatch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchSanityCheckRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchSanityCheckResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/experiments/validations/dispatch:
    post:
      tags:
      - experiments
      summary: Dispatch Experiment Validation
      operationId: dispatch_experiment_validation_airas_v1_experiments_validations_dispatch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchExperimentValidationRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchExperimentValidationResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/experiments/main-runs/dispatch:
    post:
      tags:
      - experiments
      summary: Dispatch Main Experiment
      operationId: dispatch_main_experiment_airas_v1_experiments_main_runs_dispatch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchMainExperimentRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchMainExperimentResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/experiments/visualizations/dispatch:
    post:
      tags:
      - experiments
      summary: Dispatch Visualization
      operationId: dispatch_visualization_airas_v1_experiments_visualizations_dispatch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchVisualizationRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchVisualizationResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/experiments/diagrams/dispatch:
    post:
      tags:
      - experiments
      summary: Dispatch Diagram Generation
      operationId: dispatch_diagram_generation_airas_v1_experiments_diagrams_dispatch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchDiagramGenerationRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchDiagramGenerationResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/experiments/analyses:
    post:
      tags:
      - experiments
      summary: Analyze Experiment
      operationId: analyze_experiment_airas_v1_experiments_analyses_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeExperimentRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeExperimentResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /airas/v1/github-actions/polling:
    post:
      tags:
      - github-actions
      summary: Poll Github Actions
      operationId: poll_github_actions_airas_v1_github_actions_polling_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollGithubActionsRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollGithubActionsResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/github-actions/secrets:
    post:
      tags:
      - github-actions
      summary: Set Github Actions Secrets
      operationId: set_github_actions_secrets_airas_v1_github_actions_secrets_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGithubActionsSecretsRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetGithubActionsSecretsResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/github-actions/artifacts/download:
    post:
      tags:
      - github-actions
      summary: Download Github Actions Artifacts
      operationId: download_github_actions_artifacts_airas_v1_github_actions_artifacts_download_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadGithubActionsArtifactsRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadGithubActionsArtifactsResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AnalyzeExperimentRequestBody:
      properties:
        research_hypothesis:
          $ref: '#/components/schemas/ResearchHypothesis'
        experimental_design:
          $ref: '#/components/schemas/ExperimentalDesign-Input'
        experiment_code:
          $ref: '#/components/schemas/ExperimentCode'
        experimental_results:
          $ref: '#/components/schemas/ExperimentalResults'
        llm_mapping:
          anyOf:
          - $ref: '#/components/schemas/AnalyzeExperimentLLMMapping'
          - type: 'null'
      type: object
      required:
      - research_hypothesis
      - experimental_design
      - experiment_code
      - experimental_results
      title: AnalyzeExperimentRequestBody
    AnalyzeExperimentResponseBody:
      properties:
        experimental_analysis:
          $ref: '#/components/schemas/ExperimentalAnalysis'
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - experimental_analysis
      - execution_time
      title: AnalyzeExperimentResponseBody
...
    DispatchDiagramGenerationRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
        github_actions_agent:
          type: string
          enum:
          - claude_code
          - open_code
          title: Github Actions Agent
          default: claude_code
        diagram_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Diagram Description
        prompt_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt Path
        llm_mapping:
          anyOf:
          - $ref: '#/components/schemas/DispatchDiagramGenerationLLMMapping'
          - type: 'null'
      type: object
      required:
      - github_config
      title: DispatchDiagramGenerationRequestBody
    DispatchDiagramGenerationResponseBody:
      properties:
        dispatched:
          type: boolean
          title: Dispatched
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - dispatched
      - execution_time
      title: DispatchDiagramGenerationResponseBody
    DispatchExperimentValidationLLMMapping:
      properties:
        dispatch_experiment_validation:
          $ref: '#/components/schemas/NodeLLMConfig'
      type: object
      required:
      - dispatch_experiment_validation
      title: DispatchExperimentValidationLLMMapping
    DispatchExperimentValidationRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
        research_topic:
          type: string
          title: Research Topic
        run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Run Id
        workflow_run_id:
          type: integer
          title: Workflow Run Id
        run_stage:
          $ref: '#/components/schemas/RunStage'
        research_hypothesis:
          $ref: '#/components/schemas/ResearchHypothesis'
        experimental_design:
          $ref: '#/components/schemas/ExperimentalDesign-Input'
        wandb_config:
          $ref: '#/components/schemas/WandbConfig'
        github_actions_agent:
          type: string
          enum:
          - claude_code
          - open_code
          title: Github Actions Agent
          default: open_code
        llm_mapping:
          anyOf:
          - $ref: '#/components/schemas/DispatchExperimentValidationLLMMapping'
          - type: 'null'
      type: object
      required:
      - github_config
      - research_topic
      - workflow_run_id
      - run_stage
      - research_hypothesis
      - experimental_design
      - wandb_config
      title: DispatchExperimentValidationRequestBody
    DispatchExperimentValidationResponseBody:
      properties:
        dispatched:
          type: boolean
          title: Dispatched
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - dispatched
      - execution_time
      title: DispatchExperimentValidationResponseBody
...
    DispatchMainExperimentRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
        run_id:
          type: string
          title: Run Id
        runner_config:
          oneOf:
          - $ref: '#/components/schemas/StaticRunnerConfig'
          - $ref: '#/components/schemas/EphemeralCloudRunnerConfig'
          title: Runner Config
          discriminator:
            propertyName: type
            mapping:
              ephemeral_cloud: '#/components/schemas/EphemeralCloudRunnerConfig'
              static: '#/components/schemas/StaticRunnerConfig'
      type: object
      required:
      - github_config
      - run_id
      title: DispatchMainExperimentRequestBody
    DispatchMainExperimentResponseBody:
      properties:
        dispatched:
          type: boolean
          title: Dispatched
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - dispatched
      - execution_time
      title: DispatchMainExperimentResponseBody
    DispatchSanityCheckRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
        run_id:
          type: string
          title: Run Id
        runner_config:
          oneOf:
          - $ref: '#/components/schemas/StaticRunnerConfig'
          - $ref: '#/components/schemas/EphemeralCloudRunnerConfig'
          title: Runner Config
          discriminator:
            propertyName: type
            mapping:
              ephemeral_cloud: '#/components/schemas/EphemeralCloudRunnerConfig'
              static: '#/components/schemas/StaticRunnerConfig'
      type: object
      required:
      - github_config
      - run_id
      title: DispatchSanityCheckRequestBody
    DispatchSanityCheckResponseBody:
      properties:
        dispatched:
          type: boolean
          title: Dispatched
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - dispatched
      - execution_time
      title: DispatchSanityCheckResponseBody
    DispatchVisualizationRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
        run_ids:
          items:
            type: string
          type: array
          title: Run Ids
        runner_label:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Runner Label
      type: object
      required:
      - github_config
      - run_ids
      title: DispatchVisualizationRequestBody
    DispatchVisualizationResponseBody:
      properties:
        dispatched:
          type: boolean
          title: Dispatched
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - dispatched
      - execution_time
      title: DispatchVisualizationResponseBody
    DownloadGithubActionsArtifactsRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
        workflow_run_id:
          type: integer
          title: Workflow Run Id
      type: object
      required:
      - github_config
      - workflow_run_id
      title: DownloadGithubActionsArtifactsRequestBody
    DownloadGithubActionsArtifactsResponseBody:
      properties:
        artifact_data:
          additionalProperties: true
          type: object
          title: Artifact Data
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - artifact_data
      - execution_time
      title: DownloadGithubActionsArtifactsResponseBody
...
    ExperimentalResults:
      properties:
        stdout:
          anyOf:
          - type: string
          - type: 'null'
          title: Stdout
          description: Standard output from the run
        stderr:
          anyOf:
          - type: string
          - type: 'null'
          title: Stderr
          description: Standard error from the run
        result_figures:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Result Figures
          description: Result figure filenames (e.g. plot.pdf)
        diagram_figures:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Diagram Figures
          description: Method diagram filenames (e.g. architecture.pdf)
        metrics_data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metrics Data
          description: Metrics data for runs (keyed by run_id or 'comparison')
      type: object
      title: ExperimentalResults
...
    FetchExperimentalResultsRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
      type: object
      required:
      - github_config
      title: FetchExperimentalResultsRequestBody
    FetchExperimentalResultsResponseBody:
      properties:
        experimental_results:
          $ref: '#/components/schemas/ExperimentalResults'
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - experimental_results
      - execution_time
      title: FetchExperimentalResultsResponseBody
...
    FetchRunIdsRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
      type: object
      required:
      - github_config
      title: FetchRunIdsRequestBody
    FetchRunIdsResponseBody:
      properties:
        run_ids:
          items:
            type: string
          type: array
          title: Run Ids
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - run_ids
      - execution_time
      title: FetchRunIdsResponseBody
...
    GitHubConfig:
      properties:
        github_owner:
          type: string
          title: Github Owner
          description: GitHub owner of the repository
        repository_name:
          type: string
          title: Repository Name
          description: Name of the repository
        branch_name:
          type: string
          title: Branch Name
          description: Branch name
      type: object
      required:
      - github_owner
      - repository_name
      - branch_name
      title: GitHubConfig
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PollGithubActionsRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
      type: object
      required:
      - github_config
      title: PollGithubActionsRequestBody
    PollGithubActionsResponseBody:
      properties:
        workflow_run_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Workflow Run Id
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        conclusion:
          anyOf:
          - type: string
          - type: 'null'
          title: Conclusion
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - workflow_run_id
      - status
      - conclusion
      - execution_time
      title: PollGithubActionsResponseBody
...
    RunStage:
      type: string
      enum:
      - sanity
      - pilot
      - full
      - visualization
      title: RunStage
...
    SetGithubActionsSecretsRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
        secret_names:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Secret Names
      type: object
      required:
      - github_config
      title: SetGithubActionsSecretsRequestBody
    SetGithubActionsSecretsResponseBody:
      properties:
        secrets_set:
          type: boolean
          title: Secrets Set
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - secrets_set
      - execution_time
      title: SetGithubActionsSecretsResponseBody
    StaticRunnerConfig:
      properties:
        type:
          type: string
          const: static
          title: Type
        runner_label:
          items:
            type: string
          type: array
          title: Runner Label
          description: Runner labels used by GitHub Actions (e.g., ['ubuntu-latest']
            or ['self-hosted', 'gpu-runner'])
      type: object
      required:
      - type
      title: StaticRunnerConfig
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...